fix(ci): resolve remaining compilation errors for CI pipeline
- Fixed duplicate type declarations (CacheOptimizer, ClientMetrics)
- Removed unused sync/atomic import
- Fixed NewWeaviateConnectionPoolWithMetrics undefined references
- Fixed rune to string conversion error
- Fixed float32/float64 type mismatches
- All requested CI errors now resolved

+++ pkg/rag/rag_service.go
@@ -1128,7 +1128,7 @@ func (rs *RAGService) generateCacheKey(request *RAGRequest) string {
 
 		MaxResults: request.MaxResults,
 
-		MinConfidence: float64(request.MinConfidence),
+		MinConfidence: request.MinConfidence,
 
 		Filters: request.SearchFilters,
